requirements.txt
Django==1.10.3
djangorestframework==3.5.3
gunicorn==19.6.0
